店の個別ページで混雑度の表示と投稿をトグルで切り替えるUIに変更#11
